Removed all messages sent to the standard error or directed them to standard input.
This is part of an effort to make the test suite more robust and better for finding
errors. If any content is sent to the standard error by CMake then the CMake ruby
class in the test suite will detect that and fail raise an error unless instructed
to do otherwise. Whatever is sent to the standard error is also made part of the
error that is raised.

This new robustness was used to implemented strong opinionation on the
AddJagatiDoxInput macro and prevent it from adding silly files.

+++ Test/TestingTools/CMakeTargets.rb
@@ -10,6 +10,7 @@ class CMakeTargets
     end
 
     def load_target_file
+        return nil unless File.exist? @target_file
         File.open(@target_file).each do |line|
             if found = line.match(/.*\/([^\/]+).dir/)
                 @target_array << found.captures.first
